Update include_path for PEAR modules to load
The Azure PHP SDK requires PEAR modules to function properly, but it
doesn't ship with them. The implementation requires us to load them
alongside, but we still have to tell the SDK where they are. In our case,
that means altering PHP's include_path to add our PEAR modules so the SDK
can load them without looking for them explicitly.

The directive to remove the path was part of the inital cleanup. This
commit restores it for the SSL branch, but will need to be added to all
other in-progress branches for the settings page to load and the plugin
to work correctly.