Use __newContainer__ for option value in Azure
When creating a new container from the settings or file upload pages, use a saner value to trigger the creation process than a phrase. This changes the value to '__newContainer__', which is an invalid name for a container and unlikely to be unintentionally used elsewhere.

This also allows the option to be translated inside the PHP where the form lives, instead of needing to localize the script.