Azure plugin: combine loops in putBlockBlob()
Two loops were being run one after the other on the same items with the
same counter, and nothing happening in between them. This combines the
loops to generate a block ID and upload it right away, rather than
building an array of block IDs and looping through each of them to
upload to Azure.