Azure plugin: create methods to check for containers and blobs
There's an existing blobExists method, but it checks for the presense of
metadata for a blob and not if the blob exists. It probably returns a
good enough value for checking, but the new method follows the SDK
example for checking blobs.

I extended this logic to also check for containers, so that when a blob
is checked, it'll check to make sure the container exists before trying
to query the container blindly.