Azure plugin: check before deleting large temp files
When putBlockBlob() is called on a small (<64M) file, send it to Azure
in a single block and leave the temp file. If a large file is uploaded,
set a flag, chunk it, and send the chunks. Once all is done, commit the
blob so Azure knows it's done and delete the large temp file.

This changeset adds a flag for large files so it's clear what is expected
to be deleted locally, then checks to make sure the remote file exists
before deleting the temp using core WP functions to delete the actual
temp file in addition to unsetting the file stream.