Azure plugin: sanitize the deleteBlob URL
The URL to delete an individual blob is generated for every one in the
browse tab, and is an easy place to get caught with some bad security.

This changeset switches to an absolute URL to the media-upload.php page,
sets a variable value for post_id to use in validating nonces, switches
to 'filename' as the query argument to pass the blob's filename, and
adds a nonce to the query string to use in validation.

This also adds basic l10n and a11y to the link, because it's the right
thing to do.