Azure plugin: remove empty anchor caused by typo
If a blob isn't in the whitelist of image extensions, it will show up as
a link instead of an image in the blob list. Due to a misplaced slash,
instead of the link getting closed it added an additional self-closing
anchor element directly after the link. This moves the slash, properly
closing the link and removing the empty anchor.