Azure plugin: remove broken inline styles from forms
Just setting a border width and color isn't enough; you need to set a
style on this one or it just doesn't know what to do. Margin is cool,
but is better served by moving all those inline styles that are the
exact same thing into an actual resuable container.