Azure plugin: catch the search nonce failure
After [ba0b0d7], if the nonce fails and we throw an Exception, catch it,
display the error message with flair, then die. This prevents any more
of the page loading, including the full contents of the blob. This helps
protect against unauthorized listing of the contents of a container.