Azure plugin: escape error messages sanely
The error/success messages sent after a container is created has janky
HTML that needs much love. Simply escaping it displays that markup as
plain text on the settings page, which is a bad experience. This moves
from escaping directly to using KSES to sanitize the output a bit until
we can rework the messages themselves.