Azure plugin: verify settings nonce when creating a container
Normal settings updates happen through options.php, but if a container
is created on the settings page, then the form values all get POSTed
back to the settings page itself. Before we do anything with those
values, we need to verify the nonce.

When the settings page detects that a container is created, JS changes
the action name to 'action2'. We check for this new action, combined
with the nonce created by the Settings API, before setting values.