Azure plugin: build strings sanely and escape them
Try to avoid concatenation of strings; instead use the printf family of
functions and escape the pieces.

For this error message, wp_kses_post() is preferred over esc_html()
because the error message returned from the service is in XML. If we use
esc_* functions, the raw XML is printed, but wp_kses_* strips out the
XML container and leaves the valid HTML elements like 'code' that helps
with formatting.