Azure plugin: tell PHPCS that we're checking nonces differently
PHP Code Sniffer complains if we use a passed value without immediately
verifying the nonce, but in this case the value is triggering another
logic fork that checks the nonce and permissions inside it. We can turn
off PHPCS warnings with a comment acknowledging that we're doing this on
purpose.