Azure plugin: fix broken settings check for Azure override
When is 1 not 1? When it's really "1".
Options set from a checkbox look like the number 1, but really it's the
string "1". Strict equality checks for 1 === "1" fail, because they
should, and switching to good coding standards and strict equality
exposed this bug.
Instead of trying to do string to int checks ourselves, just use the
handy 'boolval' function to do the conversion for us. If the option to
override the uploader is set in the plugin, we should do what we can to
honor that setting.