Added compatibility for WP-CLI's `media regenerate` command.
As the plugin is often used to offload all media to Azure Blob Storage, it's necessary to pull the source files back down to the web server when we're regenerating thumbnails.

As WP-CLI doesn't expose a good way to check what command is currently being run, this filter on "get_attached_file" determines if we're in a WP-CLI context, then checks the global $argv array (which is defined when PHP is operating in the 'cli' server API). If we're indeed running `media regenerate`, the filter first checks to see if the requested file exists and, if not, download it from Azure.

There are two drawbacks to this approach:

1. This is pretty wasteful with regards to bandwidth, but that's kind of the name of the game when your WP site is offloading all of its media to a CDN/storage solution.
2. This is an abuse of the "get_attached_file" filter, as we're not doing anything to *filter* the content, only causing side effects. Unfortunately, unless WP-CLI were to add a series of `do_action()` calls in the regeneration command, this is our most convenient point of injection.