Strengthen how `$upload_folder_path` is generated.
Fixes a bug with path generation on multisite, where errant `/` entries caused files to be offloaded to unnamed Azure directories, but retained the original pathing for assets, leading to broken image URLs.

This makes use of `sprintf` to generate the path for the upload folder path used for the file that should be offloaded to Azure. This improves readability of an otherwise complex set of string manipulation to ensure a consistent URL pattern at all times.